requirements.txt
Django==3.2.23
django-hosts==6.0
psycopg2-binary==2.9.9
django-extensions==3.2.3
django-cors-headers==4.3.1
django-tables2==2.7.0
hashids==1.3.1
python-slugify==8.0.1
django-widget-tweaks==1.5.0
django-ckeditor==6.7.0
mongoengine==0.27.0
Pillow==10.1.0

djangorestframework==3.13.1
django-filter==21.1
Markdown==3.5.1
drf-yasg2==1.19.4

fabric==3.2.2
invocations==3.3.0
django-debug-toolbar==4.2.0

celery==5.3.6
redis==5.0.1
django-redis==5.4.0

flake8==6.1.0
factory-boy==2.12.0
ipython==8.18.1
